feat: 이슈 컴포넌트 분할
